Revert "change fund to top off if balance below 10k xlm (#1478)"
This reverts commit 3474c736d936999961c45874ddf2fdd0b4b09f4c.
