+++ cmd/crates/soroban-test/tests/it/integration/hello_world.rs
@@ -35,7 +35,7 @@ async fn invoke() {
         .arg("fund")
         .arg("test")
         .assert()
-        .stdout(predicates::str::contains("Nothing to do."));
+        .stderr(predicates::str::contains("Account already exists"));
     sandbox
         .new_assert_cmd("keys")
         .arg("fund")
@@ -138,13 +138,6 @@ async fn invoke() {
     handles_kebab_case(sandbox, id).await;
     fetch(sandbox, id).await;
     invoke_prng_u64_in_range_test(sandbox, id).await;
-    // test fund will add when account exists
-    sandbox
-        .new_assert_cmd("keys")
-        .arg("fund")
-        .arg("test")
-        .assert()
-        .stdout(predicates::str::contains("New test balance:"));
 }
 
 fn invoke_hello_world(sandbox: &TestEnv, id: &str) {
